backend/requirements.txt
asttokens==3.0.0
awscli==1.40.21
bleach==6.2.0
blinker==1.9.0
botocore==1.38.22
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
docutils==0.19
executing==2.2.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.59.2
fsspec==2025.9.0
gunicorn==23.0.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.4
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
#pywin32 ==311
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.5
rsa==4.7.2
s3transfer==0.13.0
scikit-learn==1.7.2
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
azure-storage-blob
